# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cybersecurity Dork Library API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories with nested structure
      parameters:
        - name: flat
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{slug}:
    get:
      operationId: getCategoryBySlug
      tags: [categories]
      summary: Get category by slug with subcategories and dork count
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Category detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryDetail"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /ai/categorize:
    post:
      operationId: aiCategorizeDork
      tags: [ai]
      summary: AI categorization of a dork query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiCategorizeBody"
      responses:
        "200":
          description: Categorization result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiCategorizeResult"
  /ai/optimize:
    post:
      operationId: aiOptimizeDork
      tags: [ai]
      summary: AI optimization of a dork query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiOptimizeBody"
      responses:
        "200":
          description: Optimization result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiOptimizeResult"
  /bookmarks:
    get:
      operationId: listBookmarks
      tags: [bookmarks]
      summary: List bookmarks for current user
      responses:
        "200":
          description: Bookmark list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bookmark"
    post:
      operationId: createBookmark
      tags: [bookmarks]
      summary: Create a bookmark
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookmarkBody"
      responses:
        "201":
          description: Created bookmark
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"
  /bookmarks/{id}:
    delete:
      operationId: deleteBookmark
      tags: [bookmarks]
      summary: Delete a bookmark
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /collections:
    get:
      operationId: listCollections
      tags: [collections]
      summary: List collections
      responses:
        "200":
          description: Collection list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Collection"
    post:
      operationId: createCollection
      tags: [collections]
      summary: Create a collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCollectionBody"
      responses:
        "201":
          description: Created collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
  /collections/{id}:
    get:
      operationId: getCollection
      tags: [collections]
      summary: Get a collection with items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Collection detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionDetail"
    delete:
      operationId: deleteCollection
      tags: [collections]
      summary: Delete a collection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Dork:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        queryTemplate:
          type: string
        optimizedQuery:
          type: string
        description:
          type: string
        usageContext:
          type: string
        intentType:
          type: string
        difficulty:
          type: string
        sourceType:
          type: string
        status:
          type: string
        viewsCount:
          type: integer
        copyCount:
          type: integer
        primaryCategoryId:
          type: integer
        primaryCategory:
          $ref: "#/components/schemas/Category"
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        operators:
          type: array
          items:
            $ref: "#/components/schemas/Operator"
        platforms:
          type: array
          items:
            $ref: "#/components/schemas/Platform"
        categories:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, queryTemplate, status, createdAt]
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        icon:
          type: string
        color:
          type: string
        depthLevel:
          type: integer
        displayOrder:
          type: integer
        isActive:
          type: boolean
        parentCategoryId:
          type: integer
        dorkCount:
          type: integer
        subcategories:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        createdAt:
          type: string
      required: [id, name, slug, depthLevel]
    CategoryDetail:
      allOf:
        - $ref: "#/components/schemas/Category"
        - type: object
          properties:
            recentDorks:
              type: array
              items:
                $ref: "#/components/schemas/Dork"
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        icon:
          type: string
        color:
          type: string
        parentCategoryId:
          type: integer
        displayOrder:
          type: integer
      required: [name, slug]
    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        tagType:
          type: string
        description:
          type: string
        dorkCount:
          type: integer
      required: [id, name, slug]
...
    Operator:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        syntax:
          type: string
        description:
          type: string
        exampleUsage:
          type: string
        dorkCount:
          type: integer
      required: [id, name, syntax]
...
    Platform:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        platformType:
          type: string
        description:
          type: string
        dorkCount:
          type: integer
      required: [id, name, slug]
...
    AiCategorizeBody:
      type: object
      properties:
        query:
          type: string
        context:
          type: string
      required: [query]
    AiCategorizeResult:
      type: object
      properties:
        primaryCategory:
          type: string
        subcategories:
          type: array
          items:
            type: string
        intentType:
          type: string
        difficulty:
          type: string
        tags:
          type: array
          items:
            type: string
        platforms:
          type: array
          items:
            type: string
        operators:
          type: array
          items:
            type: string
        title:
          type: string
        description:
          type: string
      required: [primaryCategory, intentType, difficulty]
    AiOptimizeBody:
      type: object
      properties:
        query:
          type: string
        category:
          type: string
      required: [query]
    AiOptimizeResult:
      type: object
      properties:
        queryTemplate:
          type: string
        optimizedQuery:
          type: string
        title:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        relatedVariants:
          type: array
          items:
            type: string
      required: [queryTemplate, optimizedQuery, title]
    Bookmark:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        dorkId:
          type: integer
        dork:
          $ref: "#/components/schemas/Dork"
        folderName:
          type: string
        createdAt:
          type: string
      required: [id, dorkId, createdAt]
    CreateBookmarkBody:
      type: object
      properties:
        dorkId:
          type: integer
        folderName:
          type: string
      required: [dorkId]
    Collection:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        description:
          type: string
        visibility:
          type: string
        itemCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, visibility, createdAt]
    CollectionDetail:
      allOf:
        - $ref: "#/components/schemas/Collection"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/CollectionItem"
    CollectionItem:
      type: object
      properties:
        collectionId:
          type: integer
        dorkId:
          type: integer
        dork:
          $ref: "#/components/schemas/Dork"
        displayOrder:
          type: integer
        notes:
          type: string
      required: [collectionId, dorkId]
    CreateCollectionBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        visibility:
          type: string
      required: [title]
...