# lib/api-zod/src/generated/types/aiCategorizeResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cybersecurity Dork Library API
 * OpenAPI spec version: 0.1.0
 */

export interface AiCategorizeResult {
  primaryCategory: string;
  subcategories?: string[];
  intentType: string;
  difficulty: string;
  tags?: string[];
  platforms?: string[];
  operators?: string[];
  title?: string;
  description?: string;
}