# lib/api-zod/src/generated/types/createBookmarkBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cybersecurity Dork Library API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookmarkBody {
  dorkId: number;
  folderName?: string;
}