# lib/api-zod/src/generated/types/dork.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cybersecurity Dork Library API
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from "./category";
import type { Operator } from "./operator";
import type { Platform } from "./platform";
import type { Tag } from "./tag";

export interface Dork {
  id: number;
  title: string;
  queryTemplate: string;
  optimizedQuery?: string;
  description?: string;
  usageContext?: string;
  intentType?: string;
  difficulty?: string;
  sourceType?: string;
  status: string;
  viewsCount?: number;
  copyCount?: number;
  primaryCategoryId?: number;
  primaryCategory?: Category;
  tags?: Tag[];
  operators?: Operator[];
  platforms?: Platform[];
  categories?: Category[];
  createdAt: string;
  updatedAt?: string;
}