# lib/api-zod/src/generated/types/operator.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cybersecurity Dork Library API
 * OpenAPI spec version: 0.1.0
 */

export interface Operator {
  id: number;
  name: string;
  syntax: string;
  description?: string;
  exampleUsage?: string;
  dorkCount?: number;
}